Add navigation to the showcase demo
