+++ demos/showcase/package.json
@@ -12,6 +12,7 @@
         "prop-types": "^15.7.2",
         "react": "^16.8.6",
         "react-dom": "^16.8.6",
+        "react-router-dom": "^5.1.2",
         "react-scripts": "3.0.1",
         "typeface-open-sans": "^0.0.54"
     },
